precision of attributes experiments and debug groups

+++ AppleMetalLearn/Shaders/GBuffer.metal
@@ -12,7 +12,7 @@ using namespace metal;
 
 struct GBufferOut{
     float4 albedo [[color(RenderTargetAlbedo)]];
-    float4 normal [[color(RenderTargetNormal)]];
+    half4 normal [[color(RenderTargetNormal)]];
     float4 positiob [[color(RenderTargetPosition)]];
 };
 
@@ -25,7 +25,7 @@ fragment GBufferOut fragment_GBuffer(
     GBufferOut OUT;
     OUT.albedo = float4(1,1,0,1);//float4(material.baseColor, 1.0);
     //albedo.a = shadow
-    OUT.normal = float4(normalize(IN.normalWS), 1.0);
+    OUT.normal = half4(normalize(IN.normalWS), 1.0);
     OUT.positiob = float4(IN.positionWS);
     return OUT;
 }
